Further tweaks to the generation involving stairs: clumps of stairs.
This allows for rooms with clumps of two or three stairs (though not
more). This means that all of the up stairs for one level, or all of the
down stairs for one level, or a combination of up and down stairs for
the level, can be found inside the room.

This commit also disables random monster and random item generation for
the vaults layout; I'm not sure if it's possible to do the monster
generation in Lua, but I'll give it a go.